api/requirements.txt
fastapi
uvicorn[standard]
numpy
pandas
tensorflow
scikit-learn
python-multipart
pillow
